Fix properly close/disconnect events.
This stuff is a bit mindfucking 😅

The _only_ issue originally was the part where we sent `websocket.close` ASGI message to the app in response to the event `wsproto.events.CloseConnection`. As per the ASGI spec, we should send a `websocket.disconnect` event to the app, not `websocket.close`.

The other changes are reverted, as it was correctly working before.

This reverts commit de2717c895cdfc7a70b59738c7147b3357416fed.
This reverts commit 3aaefe6f271d331266b2b53ff6ca257a9e5edd9c.
This reverts commit 4af281e54d19a8e58022a419b5383eb729f5e8bb.